backend/requirements.txt

flask==3.0.0
flask-cors==4.0.0
openai==1.12.0
httpx==0.26.0
python-dotenv==1.0.0
groq==0.5.0
asyncio==3.4.3


Flask==2.3.3
Flask-CORS==4.0.0
openai==1.12.0
httpx==0.26.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0



Flask==2.3.3
Flask-CORS==4.0.0
openai==1.12.0
httpx==0.26.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==21.2.0